requirements.txt
pydantic-ai
python_dotenv